mini-services/cv-service/requirements.txt
# VisionPipeline CV Service Dependencies
# ======================================

# Core
numpy>=1.24.0
opencv-python-headless>=4.9.0

# Computer Vision Models
ultralytics>=8.1.0
deep-sort-realtime>=1.3.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0

# ReID (optional)
# torchreid>=0.2.5

# MLOps
mlflow>=2.9.0
evidently>=0.4.0

# Serving
bentoml>=1.1.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Async
aiohttp>=3.9.0
aiokafka>=0.10.0

# Task Queue
celery[redis]>=5.3.0
redis>=5.0.0

# WebSocket Server
socketio>=4.7.0
python-socketio>=5.10.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Monitoring
prometheus-client>=0.19.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# Airflow (for DAGs)
apache-airflow>=2.8.0
apache-airflow-providers-postgres>=5.6.0

# Utilities
httpx>=0.25.0
Pillow>=10.0.0
